+++ docs/Measurements.Client/ReleaseNotes/ReleaseNotes.md
@@ -1,5 +1,11 @@
 # Measurements.Client Release Notes
 
+## Version 5.0.0
+
+- Added method to get aggregated measurements for all years.
+- MeasurementAggregationByMonthsDto renamed to MeasurementAggregationByMonthDto
+- Unused properties of MeasurementAggregationByMonthDto removed
+
 ## Version 4.0.0
 
 - Aggregated Measurements endpoint routes renamed.